Moved enums to db package.
