+++ tds/modules/simulation/model.py
@@ -6,8 +6,8 @@ from typing import List, Optional
 
 from pydantic import BaseModel, Field
 
-from tds.autogen.enums import SimulationEngine, SimulationStatus, SimulationType
 from tds.db.base import TdsModel
+from tds.db.enums import SimulationEngine, SimulationStatus, SimulationType
 from tds.settings import settings
